LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-duplicate-names-list-expected.txt

CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'a' because it violates the following Content Security Policy directive: "trusted-types a b c*"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'd' because it violates the following Content Security Policy directive: "trusted-types a b c*"

PASS TrustedTypePolicyFactory and policy list in CSP.
